requirements.txt
Shapely==1.7.0
xmljson==0.2.0
thop==0.0.31.post2005241907
matplotlib==3.1.3
imgaug==0.4.0
scipy==1.4.1
p_tqdm==1.3.3
lxml==4.5.0
tqdm==4.43.0
ujson==1.35
PyYAML==5.3.1
scikit_learn==0.23.2
tensorboard==2.3.0
gdown==3.12.2
efficientnet-pytorch==0.6.3
pycapnp==1.0.0
cffi==1.14.5
pyzmq==22.0.3
crcmod==1.7.0
pymap3d==2.7.2
aenum==3.1.8
pandas